+++ json_serializable/lib/src/type_helpers/json_helper.dart
@@ -11,6 +11,9 @@ import 'package:source_gen/source_gen.dart';
 import '../shared_checkers.dart';
 import '../type_helper.dart';
 import '../utils.dart';
+import 'generic_factory_helper.dart';
+
+const _helperLambdaParam = 'value';
 
 class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
   const JsonHelper();
@@ -29,8 +32,31 @@ class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
       return null;
     }
 
-    if (context.config.explicitToJson) {
-      return '$expression${context.nullable ? '?' : ''}.toJson()';
+    final interfaceType = targetType as InterfaceType;
+
+    final toJsonArgs = <String>[];
+
+    var toJson = _toJsonMethod(interfaceType);
+
+    if (toJson != null) {
+      // Using the `declaration` here so we get the original definition –
+      // and not one with the generics already populated.
+      toJson = toJson.declaration;
+
+      toJsonArgs.addAll(
+        _helperParams(
+          context.serialize,
+          _encodeHelper,
+          interfaceType,
+          toJson.parameters.where((element) => element.isRequiredPositional),
+          toJson,
+        ),
+      );
+    }
+
+    if (context.config.explicitToJson || toJsonArgs.isNotEmpty) {
+      return '$expression${context.nullable ? '?' : ''}'
+          '.toJson(${toJsonArgs.map((a) => '$a, ').join()} )';
     }
     return expression;
   }
@@ -53,8 +79,19 @@ class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
 
     var output = expression;
     if (fromJsonCtor != null) {
-      var asCastType =
-          fromJsonCtor.parameters.singleWhere((pe) => pe.isPositional).type;
+      final positionalParams = fromJsonCtor.parameters
+          .where((element) => element.isPositional)
+          .toList();
+
+      if (positionalParams.isEmpty) {
+        throw InvalidGenerationSourceError(
+          'Expecting a `fromJson` constructor with exactly one positional '
+          'parameter. Found a constructor with 0 parameters.',
+          element: fromJsonCtor,
+        );
+      }
+
+      var asCastType = positionalParams.first.type;
 
       if (asCastType is InterfaceType) {
         final instantiated = _instantiate(asCastType as InterfaceType, type);
@@ -64,6 +101,19 @@ class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
       }
 
       output = context.deserialize(asCastType, output).toString();
+
+      final args = [
+        output,
+        ..._helperParams(
+          context.deserialize,
+          _decodeHelper,
+          targetType as InterfaceType,
+          positionalParams.skip(1),
+          fromJsonCtor,
+        ),
+      ];
+
+      output = args.join(', ');
     } else if (_annotation(context.config, type)?.createFactory == true) {
       if (context.config.anyMap) {
         output += ' as Map';
@@ -82,12 +132,95 @@ class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
   }
 }
 
+List<String> _helperParams(
+  Object Function(DartType, String) execute,
+  TypeParameterType Function(ParameterElement, Element) paramMapper,
+  InterfaceType type,
+  Iterable<ParameterElement> positionalParams,
+  Element targetElement,
+) {
+  final rest = <TypeParameterType>[];
+  for (var param in positionalParams) {
+    rest.add(paramMapper(param, targetElement));
+  }
+
+  final args = <String>[];
+
+  for (var helperArg in rest) {
+    final typeParamIndex =
+        type.element.typeParameters.indexOf(helperArg.element);
+
+    // TODO: throw here if `typeParamIndex` is -1 ?
+    final typeArg = type.typeArguments[typeParamIndex];
+    final body = execute(typeArg, _helperLambdaParam);
+    args.add('($_helperLambdaParam) => $body');
+  }
+
+  return args;
+}
+
+TypeParameterType _decodeHelper(
+  ParameterElement param,
+  Element targetElement,
+) {
+  final type = param.type;
+
+  if (type is FunctionType &&
+      type.returnType is TypeParameterType &&
+      type.normalParameterTypes.length == 1) {
+    final funcReturnType = type.returnType;
+
+    if (param.name == fromJsonForName(funcReturnType.element.name)) {
+      final funcParamType = type.normalParameterTypes.single;
+
+      if (funcParamType.isDartCoreObject || funcParamType.isDynamic) {
+        return funcReturnType as TypeParameterType;
+      }
+    }
+  }
+
+  throw InvalidGenerationSourceError(
+    'Expecting a `fromJson` constructor with exactly one positional '
+    'parameter. '
+    'The only extra parameters allowed are functions of the form '
+    '`T Function(Object) ${fromJsonForName('T')}` where `T` is a type '
+    'parameter of the target type.',
+    element: targetElement,
+  );
+}
+
+TypeParameterType _encodeHelper(
+  ParameterElement param,
+  Element targetElement,
+) {
+  final type = param.type;
+
+  if (type is FunctionType &&
+      isObjectOrDynamic(type.returnType) &&
+      type.normalParameterTypes.length == 1) {
+    final funcParamType = type.normalParameterTypes.single;
+
+    if (param.name == toJsonForName(funcParamType.element.name)) {
+      if (funcParamType is TypeParameterType) {
+        return funcParamType;
+      }
+    }
+  }
+
+  throw InvalidGenerationSourceError(
+    'Expecting a `toJson` function with no required parameters. '
+    'The only extra parameters allowed are functions of the form '
+    '`Object Function(T) toJsonT` where `T` is a type parameter of the target '
+    ' type.',
+    element: targetElement,
+  );
+}
+
 bool _canSerialize(JsonSerializable config, DartType type) {
   if (type is InterfaceType) {
     final toJsonMethod = _toJsonMethod(type);
 
     if (toJsonMethod != null) {
-      // TODO: validate there are no required parameters
       return true;
     }