+++ json_serializable/test/src/_json_serializable_test_input.dart
@@ -528,3 +528,55 @@ class OverrideGetterExampleI613Super {
 
   String get id => throw UnimplementedError();
 }
+
+@ShouldThrow(
+  'Expecting a `fromJson` constructor with exactly one positional parameter. '
+  'Found a constructor with 0 parameters.',
+  element: 'fromJson',
+)
+@JsonSerializable()
+class InvalidChildClassFromJson {
+  NoParamFromJsonCtor field;
+}
+
+class NoParamFromJsonCtor {
+  NoParamFromJsonCtor.fromJson();
+}
+
+@ShouldThrow(
+  'Expecting a `fromJson` constructor with exactly one positional parameter. '
+  'The only extra parameters allowed are functions of the form '
+  '`T Function(Object) fromJsonT` '
+  'where `T` is a type parameter of the target type.',
+  element: 'fromJson',
+)
+@JsonSerializable()
+class InvalidChildClassFromJson2 {
+  ExtraParamFromJsonCtor field;
+}
+
+class ExtraParamFromJsonCtor {
+  // ignore: avoid_unused_constructor_parameters
+  ExtraParamFromJsonCtor.fromJson(Map<String, dynamic> json, int oops);
+
+  Map<String, dynamic> toJson() => null;
+}
+
+@ShouldThrow(
+  'Expecting a `toJson` function with no required parameters. '
+  'The only extra parameters allowed are functions of the form '
+  '`Object Function(T) toJsonT` where `T` is a type parameter of the target '
+  ' type.',
+  element: 'toJson',
+)
+@JsonSerializable()
+class InvalidChildClassFromJson3 {
+  ExtraParamToJson field;
+}
+
+class ExtraParamToJson {
+  // ignore: avoid_unused_constructor_parameters
+  ExtraParamToJson.fromJson(Map<String, dynamic> json);
+
+  Map<String, dynamic> toJson(int bob) => null;
+}